requirements.txt
fastapi
uvicorn
python-dotenv
motor